+++ examples/ICMS_IPI/BlocoB.php
@@ -6,7 +6,7 @@ require_once '../../bootstrap.php';
 use NFePHP\EFD\Blocks\ICMSIPI\BlockB;
 
 //grava json das propriedades dos elementos dos blocos para a versão 017
-$bC = new BlockB('017');
+$bC = new BlockB('020');
 $std = new stdClass();
 
 foreach ($bC->elements as $key => $element) {